requirements.txt
streamlit>=1.28.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.10
chromadb>=0.4.0
python-dotenv>=1.0.0
pypdf>=3.17.0
docx2txt>=0.8